Its cobbling time

+++ core/src/ludum/mighty/square/noPlayer/GreenBase.java
@@ -11,21 +11,23 @@ import java.util.ArrayList;
  */
 public class GreenBase extends NoPlayer {
 
-	private int x, y;
+	private float x, y;
 
-	public int getX() {
+
+
+	public float getX() {
 		return x;
 	}
 
-	public void setX(int x) {
+	public void setX(float x) {
 		this.x = x;
 	}
 
-	public int getY() {
+	public float getY() {
 		return y;
 	}
 
-	public void setY(int y) {
+	public void setY(float y) {
 		this.y = y;
 	}
